Move suggestions component
