+++ src/parser/core/modules/Suggestions/Component.js
