+++ src/parser/core/modules/Suggestions/Suggestions.module.css.d.ts
